backend/requirements.txt
# Core
fastapi>=0.115.0,<1.0.0
uvicorn>=0.32.0,<1.0.0
pydantic>=2.0,<3.0.0
pydantic-settings>=2.0,<3.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.9

# LLM
openai>=1.50.0,<2.0.0
google-generativeai>=0.8.4,<1.0.0
langfuse>=2.0.0,<3.0.0

# HTTP + Retry
httpx>=0.27.0,<1.0.0
tenacity>=8.2.0,<9.0.0

# Rate limiting
slowapi>=0.1.9
